carbon_backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
supabase==2.15.0
Pillow==10.2.0
requests==2.31.0
pytest==8.3.5
pytest-django==4.8.0 